+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import { Outlet } from 'react-router-dom';
 import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
 import { Container } from 'react-bootstrap';
 
-import Header from "./components/Header/nav.jsx";
+import Header from "./components/Header/Header.jsx";
 
 const client = new ApolloClient({
   uri: '/graphql',
